Create 6.2 List class.js
Rewrite the list data structure from the exercises in Chapter 4 as a class. Give List objects their old value and rest properties, but also a toArray method and a length getter that returns the length of the list. Make fromArray a static method on the List constructor.

In order for lists to work as a class with methods, we can no longer represent the empty list as null, but have to create a special instance of our class that acts as the empty list placeholder, and compare with that instance, instead of null, when checking if we’ve reached the end of a list. Store this instance in List.empty (a static property).